requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.7.0
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11
django-cors-headers==2.4.0
django-import-export==1.1.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
jieba==0.39
lxml==4.2.6
odfpy==1.4.0
openpyxl==2.5.12
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pypiwin32==223
pytz==2018.7
pywin32==224
PyYAML==3.13
queuelib==1.5.0
requests==2.21.0
Scrapy==1.5.1
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.6.2
tablib==0.12.1
Twisted==18.9.0
unicodecsv==0.14.1
urllib3==1.24.1
w3lib==1.19.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
